move Authenticated Routes with UserSession to Userarea
