+++ routes/userarea/README.md
@@ -1,0 +1,1 @@
+Everything in this folder is authenticated and requires a valid user session.
