+++ routes/userarea/_middleware.ts
